requirements.txt
pymupdf
sentence-transformers
faiss-cpu
transformers
torch
numpy
fastapi
uvicorn
python-multipart